miopen_get_v method moved to fin_utils.py
