+++ tuna/worker_interface.py
@@ -56,7 +56,7 @@ from tuna.utils.db_utility import gen_select_objs, gen_update_query, has_attr_se
 from tuna.connection import Connection
 from tuna.utils.utility import SimpleDict
 from tuna.utils.logger import set_usr_logger
-import tuna.miopen.worker.fin_class as fc
+from tuna.miopen.worker.fin_utils import chk_miopen_env_v
 
 
 class WorkerInterface(Process):
@@ -387,7 +387,7 @@ class WorkerInterface(Process):
       raise ValueError(
           f'session rocm_v {self.dbt.session.rocm_v} does not match env rocm_v {env_rocm_v}'
       )
-    env_miopen_v: str = fc.chk_miopen_env_v(self)
+    env_miopen_v: str = chk_miopen_env_v(self)
     if self.dbt.session.miopen_v != env_miopen_v:
       raise ValueError(
           f'session miopen_v {self.dbt.session.miopen_v} does not match env miopen_v {env_miopen_v}'
